+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ##### Unreleased
 - Fixed a bug in the order of getting flags in `RegExp.prototype.flags` in the actual version of V8
 - Fixed property descriptors of some `Math` and `Number` constants
+- Added detection of NodeJS [bug](https://github.com/nodejs/node/issues/41038) in `structuredClone` that can not clone `DOMException` (just in case for future versions that will fix other issues)
 - Added NodeJS 18.3 compat data mapping
 - Updated Electron 20.0 compat data mapping